[Infra] Consolidate AuthenticationMode type into pkg/auth
Move type AuthenticationMode and its constants from
pkg/platform/kube/ingress/types.go into pkg/auth/types.go so there is
a single canonical definition. Update all call sites to import pkg/auth
instead of pkg/platform/kube/ingress for the type.
